processor/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
gunicorn>=22.0.0
meilisearch>=0.31.0,<0.41.0
psycopg2-binary>=2.9.9
redis>=5.0.0
qdrant-client>=1.12.0,<1.14.0
httpx>=0.27.0
numpy>=1.26.0
sentence-transformers>=3.0.0
transformers>=4.44.0
datasketch>=1.6.4
trafilatura>=1.6.3
langdetect>=1.0.9
beautifulsoup4>=4.12.0
spacy>=3.7.0
accelerate>=0.33.0
